+++ lib/ruby/engine_builder/builder/build_report.rb
